# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:8080",
	"basePath": "/",
	"tags": [
		{
			"name": "book-controller",
			"description": "Book Controller"
		}
	],
	"paths": {
		"/alldata": {
			"get": {
				"tags": [
					"book-controller"
				],
				"summary": "GetAllData",
				"operationId": "GetAllDataUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Book"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/book": {
			"post": {
				"tags": [
					"book-controller"
				],
				"summary": "this method is used for upsert book data into db",
				"operationId": "BookUpsertUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "book",
						"description": "book",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Book"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/del/{id}": {
			"delete": {
				"tags": [
					"book-controller"
				],
				"summary": "Delbyid",
				"operationId": "DelbyidUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		}
	},
	"definitions": {
		"Book": {
			"type": "object",
			"properties": {
				"bid": {
					"type": "integer",
					"format": "int32"
				},
				"name": {
					"type": "string"
				},
				"prise": {
					"type": "number",
					"format": "double"
				}
			},
			"title": "Book"
		}
	}
}